Update the array_practice.rb
I removed the extra array and instead ran all the methods on the existing one.
I changed the .reject to delete_if and I changed the keep to a select so as not to change the array.
I change dthe iterators from number to digit... so that I wasn't over the space limit :)

[#88972898]